doczilla-WorkingAuthentication
Updated user login so that it validates password. Moved all login functionality/views to "Home". Added "UserRole" session ID that stores the empRole field of EmployeeRole table for the employee that is logged in.
