+++ 540GPWorkingBuild/App_Start/RouteConfig.cs
@@ -16,7 +16,7 @@ namespace _540GPWorkingBuild
             routes.MapRoute(
                 name: "Default",
                 url: "{controller}/{action}/{id}",
-                defaults: new { controller = "MainLogIn", action = "Login", id = UrlParameter.Optional }
+                defaults: new { controller = "Home", action = "Login", id = UrlParameter.Optional }
             );
         }
     }
