+++ 540GPWorkingBuild/Controllers/HomeController.cs
@@ -26,5 +26,53 @@ namespace _540GPWorkingBuild.Controllers
 
             return View();
         }
-    }
-}
+
+
+
+        public ActionResult Login()
+        {
+            return View();
+        }
+
+        [HttpPost]
+        [ValidateAntiForgeryToken]
+        public ActionResult Login(Models.Employee e)
+        {
+            if (ModelState.IsValid)
+            {
+                using (Models.MusciToolkitDBEntities db = new Models.MusciToolkitDBEntities())
+                {
+                    var obj = db.Employees.Where(a => a.EmployeeID.Equals(e.EmployeeID) && a.Password.Equals(e.Password)).FirstOrDefault();
+                    if (obj != null)
+                    {
+                        Session["UserID"] = obj.EmployeeID.ToString();
+                        Session["UserName"] = obj.Password.ToString();
+                        Session["UserRole"] = obj.EmployeeRole.EmpRole.ToString();
+                        return RedirectToAction("Menu");
+                    }
+                }
+            }
+            return View(e);
+        }
+
+        public ActionResult Menu()
+        {
+            if (Session["UserID"] != null)
+            {
+                return View();
+            }
+            else
+            {
+                return RedirectToAction("Login");
+            }
+        }
+
+
+        public ActionResult Logout()
+        {
+            return View();
+        }
+
+    } // home controller
+
+} // namespace
