--- 540GPWorkingBuild/Controllers/MenuController.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-
-namespace _540GPWorkingBuild.Controllers
-{
-    public class MenuController : Controller
-    {
-        // GET: Menu
-        public ActionResult Index()
-        {
-            return View();
-        }
-    }
-}
