--- 540GPWorkingBuild/Controllers/SignOutController.cs
@@ -1,17 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
-
-namespace _540GPWorkingBuild.Controllers
-{
-    public class SignOutController : Controller
-    {
-        // GET: SignOut
-        public ActionResult Index()
-        {
-            return View();
-        }
-    }
-}
