+++ 540GPWorkingBuild/Views/Home/Logout.cshtml
@@ -1,0 +1,7 @@
+﻿
+@{
+    ViewBag.Title = "Logout";
+}
+
+<h2>Logout</h2>
+
