+++ 540GPWorkingBuild/Views/Home/Menu.cshtml
@@ -5,9 +5,10 @@
 
 <h2>Main Menu</h2>
 <p><b>This is the landing page - design to follow</b></p>
-<blockquote>POSimple is designed to aidsmall to mid-size businesses in managing sales interactions 
-    for their unique customer base. It represents a small software suitthat is capable of conducting 
-    sales transactions, supporting inventory management and customer organization, and providing sales 
+<blockquote>
+    POSimple is designed to aidsmall to mid-size businesses in managing sales interactions
+    for their unique customer base. It represents a small software suitthat is capable of conducting
+    sales transactions, supporting inventory management and customer organization, and providing sales
     reporting centered on product and employee performance.
 </blockquote>
 <h4>Software Proposal</h4>
@@ -41,4 +42,6 @@
     </ul>
 </ul>
 <p><b>Reference Mockups/SRS for full required functionality</b></p>
+
+
