--- 540GPWorkingBuild/Views/MainLogIn/UserDashBoard.cshtml
@@ -1,8 +1,0 @@
-﻿@model _540GPWorkingBuild.Models.Employee
-
-@{
-    ViewBag.Title = "UserDashBoard";
-}
-
-<h2>UserDashBoard</h2>
-
