+++ 540GPWorkingBuild/Views/Shared/_Layout.cshtml
@@ -16,7 +16,7 @@
                     <span class="icon-bar"></span>
                     <span class="icon-bar"></span>
                 </button>
-                @Html.ActionLink("Music Store Toolkit Menu", "Index", "Menu", new { area = "" }, new { @class = "navbar-brand" })
+                @Html.ActionLink("Music Store Toolkit Menu", "Menu", "Home", new { area = "" }, new { @class = "navbar-brand" })
             </div>
             <div class="navbar-collapse collapse">
                 <ul class="nav navbar-nav">
@@ -25,7 +25,7 @@
                     <li>@Html.ActionLink("Customer", "Index", "Customers")</li>
                     <li>@Html.ActionLink("Reports", "Index", "Reports")</li>
                     <li>@Html.ActionLink("Employee", "Index", "Employees")</li>
-                    <li>@Html.ActionLink("Sign Out", "Index", "SignOut")</li>
+                    <li>@Html.ActionLink("Sign Out", "Logout", "Home")</li>
                 </ul>
             </div>
         </div>
