--- 540GPWorkingBuild/Views/SignOut/Index.cshtml
@@ -1,8 +1,0 @@
-﻿
-@{
-    ViewBag.Title = "Sign Out";
-}
-
-<h2>Sign Out</h2>
-<p>Place holder - needs scripting</p>
-
