+++ 540GPWorkingBuild/Web.config
@@ -73,7 +73,7 @@
     </compilers>
   </system.codedom>
   <connectionStrings>
-    <add name="MusciToolkitDBEntities" connectionString="metadata=res://*/Models.MTKDB.csdl|res://*/Models.MTKDB.ssdl|res://*/Models.MTKDB.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=BEECHES\SQLEXPRESS;initial catalog=MusciToolkitDB;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
+    <add name="MusciToolkitDBEntities" connectionString="metadata=res://*/Models.MTKDB.csdl|res://*/Models.MTKDB.ssdl|res://*/Models.MTKDB.msl;provider=System.Data.SqlClient;provider connection string=&quot;data source=DESKTOP-N41T055\SQLEXPRESS;initial catalog=MusciToolkitDB;integrated security=True;MultipleActiveResultSets=True;App=EntityFramework&quot;" providerName="System.Data.EntityClient" />
   </connectionStrings>
   <entityFramework>
     <defaultConnectionFactory type="System.Data.Entity.Infrastructure.SqlConnectionFactory, EntityFramework" />
